cpu/arm7_common: add inlined header only def for irq_%
