+++ cpu/lpc2387/include/cpu_conf.h
@@ -38,6 +38,11 @@ extern "C" {
 /** @} */
 
 /**
+￼ * @brief   This arch uses the inlined irq API.
+￼ */
+#define IRQ_API_INLINED     (1)
+
+/**
  * @name Kernel configuration
  * @{
  */
